cache: fix integer conversion error in parsing cache_pool_max_chunks
The allocation/cache_pool_max_chunks config is read as an int32_t,
then converted to uint64_t. If the value in lvm.conf exceeds INT32_MAX,
conversion to uint64_t results in truncation followed by sign extension,
causing the original unsigned representation to be lost. For example, a
config value of 6442450944 is converted to 18446744071562067968.

This fix preserves the original u64 value from the parser. Further
adjustments should be applied to fit dm-cache's metadata limits.

Signed-off-by: Ming-Hung Tsai <mtsai@redhat.com>